pyproject.toml
[project]
name = "abnormal_detection_pretrain"
version = "0.1.0"
description = "Mask-based self-supervised pretraining for image anomaly detection"
authors = [
    { name="Chang Chun", email="goldensasuke@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "Pillow>=10.0",
    "wandb>=0.15",
]

[build-system]
requires = ["setuptools>=65.5.0","wheel"]
build-backend = "setuptools.build_meta"
